typo file renaming
