+++ credsweeper/filters/group/credit_card_number_sequence.py
